requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Werkzeug==3.1.5
SQLAlchemy==2.0.46

feedparser==6.0.12
requests==2.32.5
python-dateutil==2.9.0.post0
dnspython==2.8.0
beautifulsoup4==4.14.3
lxml==6.0.3

pyotp==2.9.0
qrcode==8.2
Pillow==12.2.0
bcrypt==5.0.0

python-nmap==0.7.1
wafw00f==2.4.2
playwright==1.58.0
playwright-stealth==2.0.2

cryptography==46.0.7
gunicorn==23.0.0
